removed the startup and updated readme file.
