+++ ServerRecording/Controllers/RecordingsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using QuickStartApi;
 using System;
 using System.IO;
 using System.Linq;
